+++ forgerock-auth/src/main/java/org/forgerock/android/auth/ServerConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 - 2023 ForgeRock. All rights reserved.
+ * Copyright (c) 2020 - 2024 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -30,6 +30,7 @@ public class ServerConfig extends NetworkConfig {
 
     public static final String API_VERSION_2_1 = "resource=2.1, protocol=1.0";
     public static final String API_VERSION_3_1 = "resource=3.1, protocol=1.0";
+    public static final String API_VERSION_1_0 = "resource=1.0";
     public static final String ACCEPT_API_VERSION = "Accept-API-Version";
 
     private String url;